+++ LibraryComponents/Classes/AndesIconsManager/AndesIcons.swift
@@ -1,0 +1,16 @@
+//
+//  AndesIcons.swift
+//  AndesUI
+//
+//  Created by Nicolas Rostan Talasimov on 2/20/20.
+//
+
+import Foundation
+
+internal struct AndesIcons {
+    static let close20 = "andes_ui_close_20"
+    static let feedbackError16 = "andes_ui_feedback_error_16"
+    static let feedbackInfo16 = "andes_ui_feedback_info_16"
+    static let feedbackSuccess16 = "andes_ui_feedback_success_16"
+    static let feedbackWarning16 = "andes_ui_feedback_warning_16"
+}