+++ LibraryComponents/Classes/AndesMessage/Type/AndesMessageTypeError.swift
@@ -9,5 +9,5 @@ import Foundation
 class AndesMessageTypeError: AndesMessageTypeProtocol {
     var primaryColor: UIColor = AndesStyleSheetManager.styleSheet.errorPrimaryColor
     var secondaryColor: UIColor = AndesStyleSheetManager.styleSheet.errorSecondaryColor
-    var icon: UIImage = UIImage(named: "andes_ui_feedback_error_16", in: AndesBundle.bundle(), compatibleWith: nil)!
+    var icon: String = AndesIcons.feedbackError16
 }