+++ LibraryComponents/Classes/Bundle/AndesBundle.swift
@@ -9,6 +9,8 @@ import Foundation
 
 public class AndesBundle {
     public static func bundle() -> Bundle {
-        return Bundle(for: self)
+        let main = Bundle(for: self)
+        let bundlePath = main.resourceURL?.appendingPathComponent("AndesUI.bundle")
+        return Bundle(url: bundlePath!)!
     }
 }